+++ lib/transpec/syntax/method_stub.rb
@@ -20,7 +20,7 @@ module Transpec
         fail 'Already replaced deprecated method, cannot allowize.' if @replaced_deprecated_method
 
         unless in_example_group_context?
-          fail NotInExampleGroupContextError.new(expression_range, "##{method_name}", '#allow')
+          fail NotInExampleGroupContextError.new(selector_range, "##{method_name}", '#allow')
         end
 
         if arg_node.type == :hash