+++ lib/transpec/syntax/should.rb
@@ -16,7 +16,7 @@ module Transpec
 
       def expectize!(negative_form = 'not_to', parenthesize_matcher_arg = true)
         unless in_example_group_context?
-          fail NotInExampleGroupContextError.new(expression_range, "##{method_name}", '#expect')
+          fail NotInExampleGroupContextError.new(selector_range, "##{method_name}", '#expect')
         end
 
         if proc_literal?(subject_node)